+++ VGEngine/Engine/source/game/textRenderSystem.cpp
@@ -52,8 +52,8 @@ void TextRenderSystem::update(std::vector<GameObject*> *gameObjects,float deltaT
 				y = base - (*glyph)->bitmap_top;
 
 				RenderSystem::updateProjection(shader);
-				shader->setUniform("unifModel", RenderSystem::modelTransform(Vector2<int>(x, y),
-					Vector2<int>((*glyph)->bitmap.width, (*glyph)->bitmap.rows), 0.0f));
+				shader->setUniform("unifModel", RenderSystem::modelTransform(Vector2<float>(x, y),
+					Vector2<float>((*glyph)->bitmap.width, (*glyph)->bitmap.rows), 0.0f));
 				shader->setUniform("unifLayer", transform->getLayer());
 				Graphics::draw(shader, text->getVertexBuffer(), text->getIndexBuffer());
